+++ src/File.php
@@ -14,4 +14,10 @@ interface File
     public function name(): Name;
     public function content(): Content;
     public function mediaType(): MediaType;
+    public function rename(Name $name): self;
+
+    /**
+     * This method called on a directory DOES NOTHING
+     */
+    public function withContent(Content $content, MediaType $mediaType = null): self;
 }